[Update]: Update Project

+++ VSA.Api/Features/Brands/DeleteBrand/DeleteBrandHandler.cs
@@ -7,57 +7,27 @@ namespace VSA.Api.Features.Brands.DeleteBrand
     public class DeleteBrandHandler : IRequestHandler<DeleteBrandCommand, DeleteBrandResponse>
     {
         private readonly AppDbContext _dbContext;
-
         private readonly Util _util;
 
-        public DeleteBrandHandler(AppDbContext dbContext)
+        public DeleteBrandHandler(AppDbContext dbContext, Util util)
         {
             _dbContext = dbContext;
-
+            _util = util;
         }
 
-        public Task<DeleteBrandResponse> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
+        public async Task<DeleteBrandResponse> Handle(DeleteBrandCommand request, CancellationToken cancellationToken)
         {
+            var brandExists = await _util.BrandIdExist(request.Id);
 
-           
-                var brandExists = _util.BrandIdExist(request.Id);
+            if (brandExists)
+            {
+                var brand = await _dbContext.Brands.FindAsync(request.Id);
 
-                var BrandResponse = new DeleteBrandResponse
+                if (brand != null)
                 {
-                    
-                };
-
-                if (brandExists)
-                {
-                    var brand =  _dbContext.Brands.FindAsync(request.Id);
                     _dbContext.Brands.Remove(brand);
-                    _dbContext.SaveChangesAsync(cancellationToken);
+                    await _dbContext.SaveChangesAsync(cancellationToken);
                 }
-                else
-                {
-                    throw new ErrorException(
-                        "DeleteBrand.NotFoundBrandId",
-                        "Not Found Brand Id! :) ");
-                }
-
-                // ... BrandResponse'i kullanarak işlemleri tamamlayabilirsiniz
-
-                return BrandResponse;
-            }
-
-
-
-            aw_util.BrandIdExist(request.Id);
-
-            var BrandResponse = new DeleteBrandResponse
-            {
-                
-            };
-
-            if(BrandId)
-            {
-                _dbContext.Remove(request.Id);
-                _dbContext.SaveChangesAsync(cancellationToken);
             }
             else
             {
@@ -66,16 +36,18 @@ namespace VSA.Api.Features.Brands.DeleteBrand
                     "Not Found Brand Id! :) ");
             }
 
-            
+            var brandResponse = new DeleteBrandResponse
+            {
+                
+            };
 
+            return brandResponse;
         }
-
-        
     }
 
-    
 }
 
-    
+
+
